+++ packages/cached_query_flutter/lib/src/query_config_flutter.dart
@@ -14,15 +14,25 @@ class QueryConfigFlutter extends QueryConfig {
   QueryConfigFlutter({
     this.refetchOnResume,
     this.refetchOnConnection,
+    @Deprecated('Use QueryConfig.storageDeserializer instead')
     Serializer? serializer,
+    Serializer? storageSerializer,
+    Serializer? storageDeserializer,
     bool? ignoreCacheDuration,
     bool? storeQuery,
     Duration? refetchDuration,
+    ShouldRefetch? shouldRefetch,
+    Duration? storageDuration,
     Duration? cacheDuration,
     bool? shouldRethrow,
   }) : super(
           ignoreCacheDuration: ignoreCacheDuration,
+          // ignore: deprecated_member_use
           serializer: serializer,
+          storageDuration: storageDuration,
+          shouldRefetch: shouldRefetch,
+          storageSerializer: storageSerializer,
+          storageDeserializer: storageDeserializer,
           storeQuery: storeQuery,
           refetchDuration: refetchDuration,
           cacheDuration: cacheDuration,